+++ app/src/main/java/com/app/miscuentas/features/inicio/InicoState.kt
@@ -3,5 +3,5 @@ package com.app.miscuentas.features.inicio
 data class InicioState(
     val huellaDigital: Boolean = false,
     val registrado: String = "",
-    val idHojaPrincipal: Int = 0
+    val idHojaPrincipal: Long = 0
 )
